rename to specify bayes
