+++ statsmodels/regression/bayes_mixed_glm.py
